Better test cases to call API
Unfortunately, our EOS node seems to not be working. The test case now
calls getBlock(), which is provided by EOS. I copied code from the eosjs
documentation, and I get a failure "Invalid block file", which is similar to the
failure when calling our test contract "Invalid block ID".

I wonder if there is something wrong with the configuration of our node.